refactor: dto 패키지 이동
- dto 패키지를 controller, service, domain과 동일한 레벨에 두었습니다.
